// frontend/src/models/apiValidationErrorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Mojica.Api
 * OpenAPI spec version: 1.0
 */
import type { ApiValidationFieldError } from "./apiValidationFieldError";

export interface ApiValidationErrorResponse {
  /** @nullable */
  code?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  errors?: ApiValidationFieldError[] | null;
}